Fix for trac ticket #4.  (Tested on DEMIS site.)
With the Identify tool on a WMS layer, now send a query to the WMS server *only* on those layers that the WMS server had already declared queryable.  This avoids the server returning an error on un-queryable layers.

Also, QgsWmsProvider::identifyAsHtml() is renamed to QgsWmsProvider::identifyAsText() to better identify what it does.  Commentary, and other dependent variable and function names have also been updated to suit.



git-svn-id: http://svn.osgeo.org/qgis/trunk@5606 c8812cc2-4d05-0410-92ff-de0c093fc19c