Real fix for trac ticket #140 (continued).
Buttons where the background color is meaningful are now rendered in classic Windows style in the Vector Properties dialog (both single value and continuous color renderers).

As part of this change the button itself is colored and not just the adjacent label.  The labels have now been removed and the buttons resized to fill the gap.

Bonus feature:  Continuous color dialog now ranges from black to white by default.


git-svn-id: http://svn.osgeo.org/qgis/trunk@5873 c8812cc2-4d05-0410-92ff-de0c093fc19c