Bug #10974 QGIS WFS Server provides too much precision
In the GetFeature request QGIS WFS Server uses the static method: QgsDoubleToString. This method returns double with a precision fixed to 17.

We do not need that much precision for coordinates. Firstly because we are not able to measure a position on earth with a lower micron precision. Secondly because it unnecessarily overload response.

I propose to add the ability to specify the precision in the QgsDoubleToString method.