[composer] Allow removing svg markers for arrows (fix #11220)
Previous behaviour was that if the start or end marker was cleared
than the old marker was drawn. This made it impossible to remove
an svg marker from the start or end of the line after one had been
assigned.

Also fix a Qt warning caused by trying to draw markers if the marker
size is set to 0.