[composer] Allow rotation for pictures set to "zoom & resize" mode
Previously rotation was only allowed for pictures set to "Zoom" mode,
which causes pictures to change size when the rotation changes (frame
size is constant). This change allows pictures set to "Zoom and resize"
mode to have rotation set. Altering the rotation when in this mode
keeps the picture size constant and changes the frame size to suit.

(fix #10640)