[FEATURE][composer] New action for toggling visibility of panels
This change adds a new action to the composer which hides or shows
all panels from the window. It's useful for maximising the space
available to interact with a composition. A shortcut (F10) and menu
item can be used to toggle this option, or via pressing 'tab' while
the main composer view is focused.