[composer] Use intersection rather than clip for transformed grids
Previous behaviour was to clip transformed grid lines to the visible
map extent. This caused problems if a map item had no border (a
solid line would be shown along the map extent), and also meant
that reprojected lines which crossed out of the map extent and back
in again did not have annotations or grid frame markings shown.

Also implement an improved logic for detecting the map frame side
for annotations.