Fix #12197 - cannot delete old visibility preset
When a preset gets applied, we try to fix it by removing irelevant stuff
(e.g. legend items that do not exist anymore, styles that do not exist).
This is however slightly complicated and leaves room for problems.

Let's simply fix the presets by assigning them the current state after
being applied - and get rid of all the patching...