osgeo4w: add a minimal exe wrapper to initialize the enviroment and load qgis_app.dll (inspired by Nathan's PR #4156)
On install each shortcut/configuration gets a own copy of the (small)
exe.  That exe is started through the batch file that originally was
used to start the app at runtime. Argument --exit will cause the wrapper
to just dump the environment to an .env file next to the copy of the
.exe.  The created shortcuts now point at those .exes instead of the
batch files.

At runtime the .exes source their .env file to get a working environment
to run QGIS (ie. qgis_app.dll).  This for example allows the application
to be pinned to the taskbar and avoids the of necessity of batch files
at runtime to start the app.

Ref: 106909fbd1b078e303e3d0041f6ca58643d6a130